C64: Better lifecycle management and 1541-II support.
