+++ BizHawk.Emulation.Cores/Computers/Commodore64/Serial/Drive1541.cs
@@ -25,6 +25,8 @@ namespace BizHawk.Emulation.Cores.Computers.Commodore64.Serial
         private bool _byteReady;
         [SaveState.SaveWithName("DriveCpuClockNumerator")]
         private readonly int _driveCpuClockNum;
+        [SaveState.SaveWithName("DriveCpuClockDenominator")]
+        private readonly int _driveCpuClockDen;
         [SaveState.SaveWithName("TrackNumber")]
         private int _trackNumber;
         [SaveState.SaveWithName("MotorEnabled")]
@@ -75,74 +77,30 @@ namespace BizHawk.Emulation.Cores.Computers.Commodore64.Serial
             Via1 = Chip6522.Create(ReadVia1PrA, ReadVia1PrB);
 
             _cpuClockNum = clockNum;
-            _driveCpuClockNum = clockDen*1000000; // 1mhz
-        }
-
-        private byte CpuPeek(ushort addr)
-        {
-            return unchecked((byte)Peek(addr));
-        }
-
-        private byte CpuRead(ushort addr)
-        {
-            return unchecked((byte) Read(addr));
-        }
-
-        private void CpuWrite(ushort addr, byte val)
-        {
-            Write(addr, val);
-        }
-
-        private bool ViaReadClock()
-        {
-            var inputClock = ReadMasterClk();
-            var outputClock = ReadDeviceClk();
-            return !(inputClock && outputClock);
-        }
-
-        private bool ViaReadData()
-        {
-            var inputData = ReadMasterData();
-            var outputData = ReadDeviceData();
-            return !(inputData && outputData);
-        }
-
-        private bool ViaReadAtn()
-        {
-            var inputAtn = ReadMasterAtn();
-            return !inputAtn;
+            _driveCpuClockNum = clockDen*16000000; // 16mhz
         }
 
         public override void ExecutePhase()
         {
-            if (_cpuClockNum > _driveCpuClockNum)
+            _ratioDifference += _driveCpuClockNum;
+            while (_ratioDifference > _cpuClockNum)
             {
-                _ratioDifference += _cpuClockNum - _driveCpuClockNum;
-                if (_ratioDifference > _cpuClockNum)
-                {
-                    _ratioDifference -= _cpuClockNum;
-                    return;
-                }
-            }
-            else if (_cpuClockNum <= _driveCpuClockNum)
-            {
-                _ratioDifference += _driveCpuClockNum - _cpuClockNum;
-                while (_ratioDifference > _driveCpuClockNum)
-                {
-                    _ratioDifference -= _driveCpuClockNum;
-                    ExecutePhaseInternal();
-                }
+                _ratioDifference -= _cpuClockNum;
+                _clocks++;
             }
             ExecutePhaseInternal();
         }
 
         private void ExecutePhaseInternal()
         {
-            Via0.Ca1 = ViaReadAtn();
-
             // clock output from 325572-01 drives CPU clock (phi0)
             ExecuteMotor();
             ExecuteFlux();
+        }
+
+        private void ExecuteSystem()
+        {
+            Via0.Ca1 = ViaReadAtn();
             Via0.ExecutePhase();
             Via1.ExecutePhase();
 
@@ -219,112 +177,6 @@ namespace BizHawk.Emulation.Cores.Computers.Commodore64.Serial
         public void RemoveMedia()
         {
             _disk = null;
-        }
-
-        public int Peek(int addr)
-        {
-            switch (addr & 0xFC00)
-            {
-                case 0x1800:
-                    return Via0.Peek(addr);
-                case 0x1C00:
-                    return Via1.Peek(addr);
-            }
-            if ((addr & 0x8000) != 0)
-                return DriveRom.Peek(addr & 0x3FFF);
-            if ((addr & 0x1F00) < 0x800)
-                return _ram[addr & 0x7FF];
-            return (addr >> 8) & 0xFF;
-        }
-
-        public int PeekVia0(int addr)
-        {
-            return Via0.Peek(addr);
-        }
-
-        public int PeekVia1(int addr)
-        {
-            return Via1.Peek(addr);
-        }
-
-        public void Poke(int addr, int val)
-        {
-            switch (addr & 0xFC00)
-            {
-                case 0x1800:
-                    Via0.Poke(addr, val);
-                    break;
-                case 0x1C00:
-                    Via1.Poke(addr, val);
-                    break;
-                default:
-                    if ((addr & 0x8000) == 0 && (addr & 0x1F00) < 0x800)
-                        _ram[addr & 0x7FF] = val & 0xFF;
-                    break;
-            }
-        }
-
-        public void PokeVia0(int addr, int val)
-        {
-            Via0.Poke(addr, val);
-        }
-
-        public void PokeVia1(int addr, int val)
-        {
-            Via1.Poke(addr, val);
-        }
-
-        public int Read(int addr)
-        {
-            switch (addr & 0xFC00)
-            {
-                case 0x1800:
-                    return Via0.Read(addr);
-                case 0x1C00:
-                    return Via1.Read(addr);
-            }
-            if ((addr & 0x8000) != 0)
-                return DriveRom.Read(addr & 0x3FFF);
-            if ((addr & 0x1F00) < 0x800)
-                return _ram[addr & 0x7FF];
-            return (addr >> 8) & 0xFF;
-        }
-
-        public void Write(int addr, int val)
-        {
-            switch (addr & 0xFC00)
-            {
-                case 0x1800:
-                    Via0.Write(addr, val);
-                    break;
-                case 0x1C00:
-                    Via1.Write(addr, val);
-                    break;
-                default:
-                    if ((addr & 0x8000) == 0 && (addr & 0x1F00) < 0x800)
-                        _ram[addr & 0x7FF] = val & 0xFF;
-                    break;
-            }
-        }
-
-        public override bool ReadDeviceClk()
-        {
-            var viaOutputClock = (Via0.DdrB & 0x08) != 0 && (Via0.PrB & 0x08) != 0;
-            return !viaOutputClock;
-        }
-
-        public override bool ReadDeviceData()
-        {
-            var viaOutputData = (Via0.DdrB & 0x02) != 0 && (Via0.PrB & 0x02) != 0;
-            var viaInputAtn = ViaReadAtn();
-            var viaOutputAtna = (Via0.DdrB & 0x10) != 0 && (Via0.PrB & 0x10) != 0;
-
-            return !(viaOutputAtna ^ viaInputAtn) && !viaOutputData;
-        }
-
-        public override bool ReadDeviceLight()
-        {
-            return _driveLightOffTime > 0;
         }
     }
 }
